:truck: Move "prop" art files into a props/ directory
