+++ src/friendo/state/fitness/read.js
@@ -2,7 +2,7 @@
 import Exercise from './exercise'
 import { left, right } from '../../../art/art-util'
 import { STATS } from '../../constants'
-import { drawOpenBook } from '../../../art/book'
+import { drawOpenBook } from '../../../art/props/book'
 
 export const ID = `state_${STATS.SIGHT}`
